Add tests verifying WC DNS records
